cloud_functions/symbol_management/requirements.txt
functions-framework==3.4.0
psycopg2-binary==2.9.7
yfinance==0.2.18
pandas==1.5.3
